Move demo-only constants from config.py to downloader.py
Demo dataset URLs, download size estimates, IMAGE_DIR, and VIDEO_DIR
were defined in the core config module but only used by demo download
code. Move them to vtsearch/datasets/downloader.py where they belong.

Also removes unused constants: AUDIO_DIR, PARAGRAPH_DIR, MEDIAS_PER_*,
IMAGES_PER_*, TEXTS_PER_CATEGORY.
