+++ vtsearch/media/video/media_type.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from vtsearch.config import UCF101_SUBSET_DOWNLOAD_SIZE_MB, VIDEO_DIR
+from vtsearch.datasets.downloader import UCF101_SUBSET_DOWNLOAD_SIZE_MB, VIDEO_DIR
 from vtsearch.media.base import (
     DemoDataset,
     MediaResponse,
